% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateFunctions.R
\name{aggregrateDataForK}
\alias{aggregrateDataForK}
\title{Aggregate a matrix of dimension M to dimension K by taking the best cut of a dendrogram}
\usage{
aggregrateDataForK(popdataraw, popdend, K, combine = matColMeans,
  simplify = TRUE)
}
\arguments{
\item{popdataraw}{A matrix}

\item{popdend}{A dendrogram relating the rows of popdataraw}

\item{K}{A height at which to cut the dendrogram}

\item{combine}{A function to combine columns of popdataraw}

\item{simplify}{Whether to simplify the labels of the dendrogram after it has been cut}
}
\description{
Cut a tree for the rows of a matrix to get K tips
Make the popdata that has this new set of rows by taking row means (or whatever combine is set to do)
}
