% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arisim.R, R/arisimsmall.R
\docType{data}
\name{arisimsmall}
\alias{arisimsmall}
\title{Small dataset for experimentation}
\format{arisimsmall contains the following items:
\itemize{
  \item mixture : the ADMIXTURE (mixture) estimate for these data
  \item data : The ChromoPainter results for these data
  \item ids : The mapping of individuals into groups 
}
#' @source \url{http://biorxiv.org/content/early/2016/07/28/066431}}
\usage{
arisimsmall

arisimsmall
}
\description{
A subset of the  \code{\link{arisim}}) data.

A subset of the  \code{\link{arisim}}) data.
}
\references{
\url{http://biorxiv.org/content/early/2016/07/28/066431}

\url{http://biorxiv.org/content/early/2016/07/28/066431}
}
\keyword{datasets}

