% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateFunctions.R
\name{dendrogramHeights}
\alias{dendrogramHeights}
\title{Find all of the unique heights of a dendrogram at which it can be cut}
\usage{
dendrogramHeights(tdend)
}
\arguments{
\item{tdend}{A dendrogram object}
}
\description{
Obtain all cuts of a dendrogram
}

