% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{idsToList}
\alias{idsToList}
\title{Convert an ID table into a population list}
\usage{
idsToList(ids)
}
\arguments{
\item{ids}{A data frame containing 2 or more columns. The first are individual identifiers, the second are population identifiers. Any further columns are ignored. NB: Chromopainter excludes individuals in the third column that have a 0; you are advised to do this before passing to this function.}
}
\value{
A list of populations, named  by the population identifiers. Each is a chartacter vector of the individual identifiers in each population.
}
\description{
Takes a data frame converting individual identifiers into cluster ids, and returns this same structure. The cluster ids (and individual ids within each cluster id) appear in the same order they appeared in the data.
}
