% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateFunctions.R
\name{MakeColorWYRPB}
\alias{MakeColorWYRPB}
\title{Make a colour palette moving from White to Yellow to Red to Purple to Blue}
\usage{
MakeColorWYRPB(colby = 0.05)
}
\arguments{
\item{colby}{The step size of the difference between successive colours. Each colour transition is characterised by moving one or move RGB colours from 0 to 1 (or vice versa) in steps of colby. Default: 0.05}
}
\description{
Make a colour palette moving from White to Yellow to Red to Purple to Blue.
}

