% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateFunctions.R
\name{MakeColorYRP}
\alias{MakeColorYRP}
\title{#' @title Make a colour palette moving from Yellow to Red to Purple}
\usage{
MakeColorYRP(colby = 0.05, final = NULL)
}
\arguments{
\item{colby}{The step size of the difference between successive colours. Each colour transition is characterised by moving one or move RGB colours from 0 to 1 (or vice versa) in steps of colby. Default: 0.05}

\item{final}{A colour to place at the end, expressed as an RGB numeric vector. Default: NULL, meaning place no additional colour at the end.}
}
\description{
Make a colour palette moving from Yellow to Red to Purple.
}
