% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{matColMeans}
\alias{matColMeans}
\title{Aggregate a matrix by taking column means}
\usage{
matColMeans(mat, poplist)
}
\arguments{
\item{mat}{A matrix}

\item{poplist}{list of which members of the original mat are in each of the new columns}
}
\description{
Take means over columns in a matrix, by grouping all columns listed in poplist
e.g. if mat is M*N matrix and poplist is length K, returns a M*K matrix
the names of poplist are used to assign names to the returned matrix
}
