% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateFunctions.R
\name{matColSums}
\alias{matColSums}
\title{Aggregate a matrix by taking column sums}
\usage{
matColSums(mat, poplist)
}
\arguments{
\item{mat}{A matrix of dimension M by N}

\item{poplist}{list of which members (column identifiers) of the original mat are in each of the new columns, as returned by \code{\link{idsToList}}.}
}
\value{
A matrix on dimension M by K.
}
\description{
Sums over columns in a matrix, by grouping all columns listed in poplist
e.g. if mat is M*N matrix and poplist is length K, returns a M*K matrix
the names of poplist are used to assign names to the returned matrix
}
