% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{mixPainter}
\alias{mixPainter}
\title{Find and test mixPainter binaries}
\usage{
mixPainter()
}
\value{
a character string of the location of mixPainter
}
\description{
badMIXTURE uses chromosome painting to test whether a given mixture is
a good explanation of the data. mixPainter is a version of
finestructure/ChromoPainter that makes this simple.

This function gives you the command-line location of a version of mixPainter
on your system. It looks for "mixPainter" in your path, or asks you to download
the mixPainter zip package and put the results somewhere it can find.
Note that it
is possible that no versions work on your system; currently we do support:
* linux (gcc 5.3.0, compiled on Scientific Linux release 6.4)
* Mac (g++-5 5.2.0, OSX Sierra 10.12)
We will make every attempt to provide a working version for other mac or
linux versions but can make no guarantee that we can get it working for you.
In that case you would be able to script your own interface with chromopainter.

It will also print the version information of mixPainter.
}
