% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixturePlot.R
\name{mixturePlot}
\alias{mixturePlot}
\title{Plot a mixture fit vs the data.}
\usage{
mixturePlot(adc, show = c("Admixture", "Cluster", "Residual"),
  mainname = "Admixture", uselayout = TRUE, popy = 0.5,
  height.cluster = 0.2, cex = 1, cex.main = 1, cex.axis = 1,
  cex.poplab = 0.6, cex.residlab = 0.6, cex.residleg = 0.6,
  cex.names = 0.2, cex.ancestral = 1, marx = c(3, 1),
  mary.admixture = c(2, 3), mary.ancestral = c(1, 2), mary.cluster = c(1,
  1), mary.residual = c(1, 1), mary.diff = c(2, 2), mary.raw = c(1, 1),
  omni.lwd = 2, omni.ylim = NULL, line.above = 0, line.below = 0,
  xlim = NULL, showtitles = TRUE, ancestralshow = NULL,
  ancestralspace = 0.5, residual.max = NULL, residual.scale = 0.1,
  residual.showscale = TRUE, ancestral.mean = TRUE,
  ancestral.names = NULL, ancestral.shownames = TRUE, cluster.srt = 0,
  cluster.names = TRUE, resid.cols = MakeColorRYWGB(), gpref = NULL)
}
\arguments{
\item{adc}{A mixture solution with data, as returned by \code{\link{compareMixtureToData}}) or \code{\link{compareMixtureToDataDirect}}).}

\item{show}{A character vector of plot types to create. See plots below.}

\item{mainname}{Title for the main plot}

\item{uselayout}{Whether we call layout, or allow the user to set up the plot environment. Default: TRUE}

\item{popy}{height for the cluster names; repeated (default: 0.5)}

\item{height.cluster}{Height of the "cluster" plot relative to the others}

\item{cex}{Overall scale; see par. Default: 1.0}

\item{cex.main}{As described in par. Default: 1.0}

\item{cex.axis}{As described in par. Default: 1.0}

\item{cex.poplab}{cex for the population labels. Default: 0.6}

\item{cex.residlab}{cex for the residuals. Default: 0.6}

\item{cex.residleg}{cex for the legend in residual plots. Default: 0.6}

\item{cex.names}{cex for the names ...? Default: 0.2}

\item{cex.ancestral}{cex for the ancestral ...? Default: 1.0}

\item{marx}{Margin in the x direction. Default: c(3,1)}

\item{mary.admixture}{Margin in the y direction of the admixture plot. Default: c(2,3)}

\item{mary.ancestral}{Margin in the y direction of the ancestral plot. Default: c(1,2)}

\item{mary.cluster}{Margin in the y direction of the cluster plot. Default: c(1,1)}

\item{mary.residual}{Margin in the y direction of the residual plot. Default: c(1,1)}

\item{mary.diff}{Margin in the y direction of the difference, diff & omni plots. Default: c(2,2)}

\item{mary.raw}{Margin in the y direction of the raw plot. Default: c(1,1)}

\item{omni.lwd}{Line width for the omni plot. Default: 2.0}

\item{omni.ylim}{Limits of the omni plot. default: NULL, meaning the range of the data}

\item{line.above}{Line where text is written above the plots with mtext; Default: 0}

\item{line.below}{Line where text is written below the plots with mtext; Default: 0}

\item{xlim}{xlim for the plots; all plots are restricted to this range. Note that you must calculate the spaces between populations; see below. (Default: NULL, meaning the range of the data)}

\item{showtitles}{whether to plot any titles. Default: TRUE}

\item{ancestralshow}{A vector of which of the ancestral populations to show. Default: NULL, meaning all of them.}

\item{ancestralspace}{Spacing between ancestral vectors (they are each of width 1). Default: 0.5}

\item{residual.max}{Maximum value for the resiidual plots. Default:NULL, meaning the range of the data.}

\item{residual.scale}{Scale of the residuals in the AdmixtureResidualMatrix, as compared to the ..? Defualt: 0.1.}

\item{residual.showscale}{Whether to show the scale in the residual matrix plot. Default: TRUE}

\item{ancestral.mean}{Whether to show the global mean in the Ancestral plot. Default: TRUE}

\item{ancestral.names}{A vector of names for the ancestral populations in the Ancestral plot. Default: NULL, meaning label them A1..K.}

\item{ancestral.shownames}{Whether to show the names in the ancestral plot. Default: TRUE.}

\item{cluster.srt}{String rotation in the Cluster plot, as passed to par(srt). Default: 0.}

\item{cluster.names}{Whether to show the names in the Cluster plot. Default: TRUE.}

\item{resid.cols}{Colours for the Residual Matrix. Default: \code{\link{MakeColorRYWGB}})(), which makes one direction Yellow->Red, and the other Green->Blue, with white around zero.}

\item{gpref}{A list of the above preferences, passed to the helper functions that do the plotting}
}
\value{
gpref A list of the graphical preferences that have been specified, so you can call this function again and get consistent output.
}
\description{
This function takes a mixture solution as returned by \code{\link{compareMixtureToData}}) or \code{\link{compareMixtureToDataDirect}}) and plots it in a variety of ways.

Focus your attention on only the first two arguments: \code{adc} which is the processed mixture solution) and \code{show} which is how you choose the plots to display. The remaining arguments affect only the presentation of the plot.

It is essential to understand that there are two key structures being explored simultaneously in these plots. The first is the P-dimensional clustering of the data, which defines the similarities; each of the N data points has a similarity to these P clusters. A representation in this space is called a palette. The second is the K-dimensional set of admixture weights. Each of the K "ancestral" or latent variables also has a P dimensional palette. Further, each of the N data points has a K dimensional "admixture" breakdown, seen as a mixture of the K ancestral palettes.

The variety of plots available each explore different aspects of the fit of the mixture solution to the data.  We assume that the individuals are ordered by cluster and plot clusters as separated. These must match the palette in order.

A variety of plot types for "show" are available:
\itemize{
  \item Admixture : the Admixture profile, shown via the classic barplot. 
  \item Cluster : The assignment of individuals into the P clusters.
  \item Raw : The palettes, as observed in the data.
  \item Pred : The predicted palettes.
  \item Ancestral : The palettes of the ancestral populations.... See ancestralshow.
  \item Residual : The matrix of residuals. See resid.cols.
  \item Same : The part of the palette that was predicted and is present.
  \item Omni : A visualisation of the raw palette and its fit to the data. The palette is split into three parts: a) the predicted palette that is present in the data, shown immediately above the x-axis up to the black line. b) the additional palette that is present in the data but missing in the prediction, shown above the black line up to 1. c) the predicted palette that is not present in the data; this is shown beneath the x axis. Therefore everything shown above the x-axis represents the data, and everything below the black line represents the prediction.
}

There are also some other plots which are more experimental:
\itemize{
  \item Over : A comparison of the predicted palettes to the observed palette; the predicted palette is shown on top, with "over-prediction" shown in black. The "missing" part of the prediction is show at the bottom.
  \item Over2 : As "Over", but without "blacking out" the section of the prediction that was overpredicted.
  \item Under : 
  \item Under2 :
}

If you want to produce high quality plots, you may wish to control the plotting very carefully. It is still helpful to use these functions as they ensure that the x-axis aligns; however you might want to display your own annotation.

This is possible by calling with a single "show" argument, and setting  layout=FALSE. Then the plot will be drawn in the next available plotting area. You can disable all text overlay and add this yourself. See examples.
}
\section{Plot types}{
NA
}

\section{Advanced usage}{
NA
}

\examples{
\dontrun{
## Straightforward example:
data(arisim)
adm<-compareMixtureToData(arisim$mixture,arisim$data,arisim$ids)
gpref=mixturePlot(adm)

## Example showing individual names clearly
data(arisimsmall)
adm<-compareMixtureToData(arisimsmall$mixture,arisimsmall$data,arisimsmall$ids)

gpref=mixturePlot(adm,
                  c("Admixture","Cluster","Raw","Residual"),
                  cex.names=0.5,residual.scale=0.2,height.cluster=.4)

## Complex example using layout and data subsetting
data(arisimsmall)
adm<-compareMixtureToData(arisimsmall$mixture,arisimsmall$data,arisimsmall$ids)
adm2=adm
adm2$mycols2[]="lightgrey"
adm2$mycols2[4:7]=c("darkred","orange","green","blue")

layout(matrix(c(1,2,3,4,5,5,5,5),nrow=4,ncol=2),heights=c(1,0.4,1,1))
gpref=mixturePlot(adm2,
                  c("Admixture","Cluster","Raw","Residual"),
                  mainname=paste0("Admixture, Marginalization, Direct"),
                  mary.admixture=c(0,1),
                  cex.names=0,
            ancestralshow=c(1,2,4), 
            residual.max=NULL,
            residual.scale=0.2,
            residual.showscale=TRUE, 
            ancestral.mean=FALSE,
            xlim=c(24,56),
            uselayout=FALSE)
gpref=mixturePlot(adm2,"Ancestral",gpref=gpref,uselayout=FALSE)
}
}
\keyword{admixture}
\keyword{chromopainter}
\keyword{finestructure}
\keyword{mixture}
\keyword{structure}
