% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMixtureToData.R
\name{oneByOnePrediction}
\alias{oneByOnePrediction}
\title{Obtain a prediction for the ancestral clusters, removing drift}
\usage{
oneByOnePrediction(mix, data.NbyP, coancestry.KbyP, pred.NbyPatK,
  tdend = NULL, clusterlist = NULL, ancestral = "solve")
}
\arguments{
\item{mix}{A proposed mixture solution of dimension N by K; for example, as generated by STRUCTURE or ADMIXTURE.}

\item{data.NbyP}{A matrix containing the data of dimension N by P, for example, as generated by ChromoPainter. the data must reflect similarity to P clusterings of the data.}

\item{coancestry.KbyP}{A matrix of the cluster averaged data}

\item{pred.NbyPatK}{A matrix of the cluster predictions}

\item{tdend}{A dendrohram of the populations, used for ordering. Default: NULL, meaning use clusterlist}

\item{clusterlist}{A list of populations, as described in \code{\link{compareMixtureToData}}. Default: NULL, meaning use tdend}

\item{ancestral}{Ancestral model, as described in \code{\link{compareMixtureToData}}. Default: "solve"}
}
\value{
A matrix of dimension P by P
}
\description{
Obtain the prediction of a mixture for each cluster (a P by P matrix), and return a matrix of this dimension but with replace the data for that cluster by its prediction, so that the diagonal is forced to match the prediction. This removes "genetic drift" specific to a given population.
}

