% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateFunctions.R
\name{popDendRelabelMembers}
\alias{popDendRelabelMembers}
\title{Relabel a dendrogram based on IDs}
\usage{
popDendRelabelMembers(popdend, ids, relabel = function(x) x)
}
\arguments{
\item{popdend}{A dendrogram}

\item{ids}{An N by 3 data frame consisting of: column 1: row names (for both the data and the mix). column 2: the cluster menbership that created the groups in dataraw (with the column names in dataraw as the values). column 3: inclusion (0 for absent, 1 for present).}

\item{relabel}{A function that takes the names in ids[,1] and returns the names that will be found in popdend.}
}
\description{
Look at each tip of popdend, and find all individuals in the ids structure that have this population. Replace labels with a ; separated character string of the individuals in that population
}
