% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{readChunkcounts}
\alias{readChunkcounts}
\title{Load a chunkcounts matrix from chromopainter}
\usage{
readChunkcounts(filename, skip = 1)
}
\arguments{
\item{filename}{Name of the chunkcounts file to be loaded}

\item{skip}{number of lines to skip at the top of the file. Set to 1 if there is a CFACTOR header.}
}
\value{
A matrix of normalised chunkcounts
}
\description{
Read chromopainter output and present it in the form that badMIXTURE needs; i.e. a matrix normalised to have rowsums = 1, which is the default behaviour
}
\examples{
\dontrun{
recent_ariL=readChunkcounts((system.file("extdata",
"Recent_admix.pruned_cp_pop_linked.chunkcounts.out", package = "badMIXTURE")))
}
}
\seealso{
readQ
}
