% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{readQ}
\alias{readQ}
\title{Load an ADMIXTURE Q file}
\usage{
readQ(filename, ids, poporder = NULL, thresh = 0.95)
}
\arguments{
\item{filename}{Name of the ADMIXTURE Q file to read}

\item{ids}{An ids dataframe; individual ids in column 1, population ids in column 2}

\item{poporder}{The order that we we want populations to be displayed in. Should correspond to the unique labels in the ids file. No reordering done if not provided}

\item{thresh}{A threshold for counting an individual as a member of a population, for use in reordering columns of Q to match the order in poporder}
}
\value{
A matrix of normalised chunkcounts
}
\description{
Read ADMIXTURE output, and assign row names to it. Optionally, reorder it (if poporder is specified)
}
\examples{
\dontrun{
## Admixture Q matrices don't have row names. Read an ID file for this:
recent_ariids=read.table(system.file("extdata",
  "Recent_admix.ids", package = "badMIXTURE"),as.is=T)
## Now read the Q matrix
recent_ariQ=readQ(system.file("extdata",
  "Recent_admix.pruned.11.Q", package = "badMIXTURE"),recent_ariids)

## Now we want to reorder the individuals, specifying the order that
we would like our populations to appear
poporder=paste0("Pop",c(7,5,6,13,4,8,9,11:12,1:3))
recent_ariQ=readQ(system.file("extdata",
  "Recent_admix.pruned.11.Q", package = "badMIXTURE"),recent_ariids,poporder)
}
}
\seealso{
reorderQ
}
