% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateFunctions.R
\name{rgbDistCols}
\alias{rgbDistCols}
\title{Obtain a colour palette for a matrix using multi-dimensional scaling}
\usage{
rgbDistCols(mydata, colorder = c(1, 2, 3, 4), colmax = 0.8,
  alphamin = 0.5)
}
\arguments{
\item{mydata}{data for which colours are required; each row is an observation}

\item{colorder}{Mapping of MDS directions to RGBalpha channels;allows relabelling of colour directions and changing polarity (if -ve)}

\item{colmax}{the maximum a colour can be (for preventing white)}

\item{alphamin}{the minimum the alpha can be (for preventing colours from being too faded)}
}
\description{
Makes a colour for each row of the data by embedding it in 4 dimensions, "r,g,b,alpha"
}
