% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateFunctions.R
\name{uniqueCutHeights}
\alias{uniqueCutHeights}
\title{Find all of the unique heights of a dendrogram at which it can be cut}
\usage{
uniqueCutHeights(tdend)
}
\arguments{
\item{tdend}{A dendrogram object}
}
\description{
Obtain all cuts of a dendrogram, return this in a useful format
}
